More session switch improvements
First, `.tmux.conf`: The command now prints the query as well as the result. It
prints the query first and the the result. If there is no result, it prints
only the query. This is useful because we can always get the right thing. I
would have sorta preferred alt-enter to work, but this is smooth enough. So,
now, after piping the result of `fzf-tmux` through to `tail` we onlt get the
last line. If this is a session, the session is the last line, so we switch.
If this is new, the query is the only line, so we have that name and create
the session.

Second, `tmux-sessions`: If this the first time tmux started from a new
console, there is no "previous session", so that returns empty. We should
detect an empty previous and just let the whole thing return unchanged. This
means `<Prefix> m` works, even on the first entrance of tmux.